libm: ref: sinpi: warning message cleanup
Chages to cleanup warning messages.

AMD-Internal: [SWLCSG-545]

+++ src/ref/sinpi.c
@@ -94,10 +94,12 @@ static inline double cos_piby4(double x, double xx)
 double FN_PROTOTYPE_REF(sinpi)(double x)
 {
     double r, dx, xsgn;
-    long long ux;
+    unsigned long long ux;
     const double pi = 3.1415926535897932384626433832795;
+
+
     GET_BITS_DP64(x, ux);
-		ux = ux & ~SIGNBIT_DP64;
+    ux = ux & ~SIGNBIT_DP64;
 
     // Handle +-Inf and NaN
     if (ux >= PINFBITPATT_DP64)
@@ -126,8 +128,8 @@ double FN_PROTOTYPE_REF(sinpi)(double x)
     }
 
     // Remaining cases
-    ux = (long)dx;
-    r = dx - ux;
+    ux = (unsigned long)dx;
+    r = dx - (double)ux;
     xsgn = (x > 0.0 ? 1.0 : -1.0) * (ux & 0x1 ? -1.0 : 1.0);
 
     if (r == 0.0)
